Renamed module 'modeling.py' to 'math_modeling.py' Renamed class LinearModel to MathModel Renamed all instances of MathModel (before LinearModel) to math_model
