+++ flixOpt/components.py
@@ -11,7 +11,7 @@ from typing import Union, Optional, Literal
 
 from flixOpt.elements import Flow, Component, MediumCollection, Global
 from flixOpt.core import Skalar, Numeric, Numeric_TS, TimeSeries, effect_values_to_ts
-from flixOpt.modeling import VariableTS, Equation
+from flixOpt.math_modeling import VariableTS, Equation
 from flixOpt.structure import SystemModel
 from flixOpt.features import FeatureLinearSegmentSet, FeatureInvest, FeatureAvoidFlowsAtOnce
 from flixOpt.flixBasicsPublic import InvestParameters, TimeSeriesRaw
