+++ flixOpt/elements.py
@@ -11,7 +11,7 @@ import logging
 
 import numpy as np
 
-from flixOpt.modeling import Variable, VariableTS, Equation
+from flixOpt.math_modeling import Variable, VariableTS, Equation
 from flixOpt.core import TimeSeries, Numeric, Numeric_TS, Skalar, as_effect_dict, as_effect_dict_with_ts
 from flixOpt.flixBasicsPublic import InvestParameters
 from flixOpt.structure import Element, SystemModel
