+++ flixOpt/features.py
@@ -15,7 +15,7 @@ import numpy as np
 from flixOpt.structure import Element, SystemModel
 from flixOpt.elements import Flow, EffectTypeDict, Global
 from flixOpt.core import TimeSeries, Skalar, Numeric, Numeric_TS, as_effect_dict
-from flixOpt.modeling import Variable, VariableTS, Equation
+from flixOpt.math_modeling import Variable, VariableTS, Equation
 from flixOpt.flixBasicsPublic import InvestParameters
 import flixOpt.flixOptHelperFcts as helpers
